Tests all working, fixed up cppnToBody to verify against c# code.
We have working cppn to body, and working genome to body.

Using neatjs we can now create bodies using genomes inside of the browser!

Therefore we can combine any individuals in the chart, that seems cool! We
can also start coding up evolution to allow for interactively evolving
iesor creatures in the browser. This is gonna be fun... :)